Confirm Delete
In the updated code snippet, the showDeleteStudentConfirmation method is used to display the confirmation message to the user by returning the confirm-delete view template. The id path variable is added to the model to set the th:action attribute of the form in the confirm-delete.html page.

The deleteStudentInformation method is updated to handle the confirmation message by checking the value of the confirm parameter. If the value is yes, the student information is deleted from the database else redirected to list page